Fix subscriptions and drop iterall
Async iterators/generators are supported since node 10.3
See here https://node.green/#ES2018-features-Asynchronous-Iterators-for-await-of-loops
We can drop now iterall.

To not break anything covered subscriptions with tests.
And found "stop" does not work because invoked twice and
fails the second time caz does not find item in this.clients map.